refactor(rum-explorer): move (some parts of) cwv timeline chart into separate class
